Add file to rustlib IDF target, deploy header file with library
If a component is registered with no files, the ESP-IDF creates a CONFIG_ONLY
component (and an INTERFACE library target). This can not be used to build
library files.

Trick the IDF into 'building' something by specifying a header file as a source file.

Specify the include directory so that component users automatically include the
Rust crate's C API. CMake's library target encapsulate both linker flags for
the library binary and preprocessor flags for the library API, so it should be
possible to specify both the Rust binary and C API in a single CMake target.