Link rustlib library against cargo artifact and clib component
The rustlib component privately requires clib.
This should force clib to be built before rustlib.
It also exposes the idf::clib alias target within the rustlib component's
CMakeLists.txt file.

Specify that the Rust ExernalProject target depends on the clib alias target.
This forces clib to be built before the Rust crate.

Collect the Rust crate's archive and its dependencies in an IMPORTED library
target using the IDF function add_prebuilt_library(). This function captures
linker flags for all of the library files into a single CMake target. Note that
this library target is IMPORTED so there is no rule to build it, and that it is
not an IDF component so it can not be required by other components.

Capture the Rust library and its dependencies in the IDF rustlib component using
target_link_libraries(). This allows other IDF components that require the rustlib
component to inherit the correct linker flags.